refactor: generalized RoPE
